app/requirements/dev.txt
-r base.txt
pytest-django==4.5.2
factory-boy==3.2.0
pytest-factoryboy==2.5.0
pytest-mock==3.8.2
Faker==13.15.0
time-machine== 2.8.2